requirements.txt
streamlit
yfinance
pandas
matplotlib
google-generativeai